+++ stencil/main.py
@@ -8,12 +8,6 @@ for entry_point in iter_entry_points('stencils'):
     stencils[entry_point.name] = entry_point.load()
 
 
-def print_list(option, opt, value, parser):
-    for name in sorted(stencils):
-        print(name)
-    sys.exit(0)
-
-
 def run():
     parser = argparse.ArgumentParser()
     parser.add_argument('-d', '--use-defaults', dest='use_defaults',
@@ -21,10 +15,11 @@ def run():
                         help="don't ask for variables with defaults if set")
     subparsers = parser.add_subparsers(title='available stencils',
                                        metavar='<stencil>')
-    for name, stencil in stencils.items():
-        stencil.add_to_subparsers(name, subparsers)
+    for name in sorted(stencils):
+        stencils[name].add_to_subparsers(name, subparsers)
     args = parser.parse_args()
     args.func(args)
+
 
 if __name__ == '__main__':
     run()