+++ HISTORY.rst
@@ -1,6 +1,11 @@
 History
 -------
 
+next release
+^^^^^^^^^^^^
+
+* Fix stencils order in command output
+
 0.2 (2011-08-05)
 ^^^^^^^^^^^^^^^^