+++ HISTORY.rst
@@ -1,6 +1,12 @@
 History
 -------
 
+0.2.1 (2011-08-10)
+^^^^^^^^^^^^^^^^^^
+
+* Better handling of target directory creation
+* Fix stencils order in command output
+
 0.2 (2011-08-05)
 ^^^^^^^^^^^^^^^^